requirements.txt
seaborn
biopython